+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unpublished] 
+
+### Fixed
+
+- traceback now correctly retrieve source for the `__main__` script when current working dir has change
+- tracebacks no longer slow down when source files can not be found.
+
 ## [9.8.2] - 2021-01-15
 
 ### Fixed